modify struct
