+++ .gitignore
@@ -1,4 +1,5 @@
 
+.idea
 vendor/
 article_clean
 _*
