+++ internal/rest/article_test.go
@@ -34,7 +34,7 @@ func TestFetch(t *testing.T) {
 
 	e := echo.New()
 	req, err := http.NewRequestWithContext(context.TODO(),
-		echo.GET, "/article?num=1&cursor="+cursor, strings.NewReader(""))
+		echo.GET, "/service?num=1&cursor="+cursor, strings.NewReader(""))
 	assert.NoError(t, err)
 
 	rec := httptest.NewRecorder()
@@ -58,7 +58,7 @@ func TestFetchError(t *testing.T) {
 	mockUCase.On("Fetch", mock.Anything, cursor, int64(num)).Return(nil, "", domain.ErrInternalServerError)
 
 	e := echo.New()
-	req, err := http.NewRequestWithContext(context.TODO(), echo.GET, "/article?num=1&cursor="+cursor, strings.NewReader(""))
+	req, err := http.NewRequestWithContext(context.TODO(), echo.GET, "/service?num=1&cursor="+cursor, strings.NewReader(""))
 	assert.NoError(t, err)
 
 	rec := httptest.NewRecorder()
@@ -87,12 +87,12 @@ func TestGetByID(t *testing.T) {
 	mockUCase.On("GetByID", mock.Anything, int64(num)).Return(mockArticle, nil)
 
 	e := echo.New()
-	req, err := http.NewRequestWithContext(context.TODO(), echo.GET, "/article/"+strconv.Itoa(num), strings.NewReader(""))
+	req, err := http.NewRequestWithContext(context.TODO(), echo.GET, "/service/"+strconv.Itoa(num), strings.NewReader(""))
 	assert.NoError(t, err)
 
 	rec := httptest.NewRecorder()
 	c := e.NewContext(req, rec)
-	c.SetPath("article/:id")
+	c.SetPath("service/:id")
 	c.SetParamNames("id")
 	c.SetParamValues(strconv.Itoa(num))
 	handler := rest.ArticleHandler{
@@ -129,7 +129,7 @@ func TestStore(t *testing.T) {
 
 	rec := httptest.NewRecorder()
 	c := e.NewContext(req, rec)
-	c.SetPath("/article")
+	c.SetPath("/service")
 
 	handler := rest.ArticleHandler{
 		Service: mockUCase,
@@ -153,12 +153,12 @@ func TestDelete(t *testing.T) {
 	mockUCase.On("Delete", mock.Anything, int64(num)).Return(nil)
 
 	e := echo.New()
-	req, err := http.NewRequestWithContext(context.TODO(), echo.DELETE, "/article/"+strconv.Itoa(num), strings.NewReader(""))
+	req, err := http.NewRequestWithContext(context.TODO(), echo.DELETE, "/service/"+strconv.Itoa(num), strings.NewReader(""))
 	assert.NoError(t, err)
 
 	rec := httptest.NewRecorder()
 	c := e.NewContext(req, rec)
-	c.SetPath("article/:id")
+	c.SetPath("service/:id")
 	c.SetParamNames("id")
 	c.SetParamValues(strconv.Itoa(num))
 	handler := rest.ArticleHandler{
