+++ service/article_services_impl.go
@@ -1,46 +1,13 @@
-package article
+package service
 
 import (
 	"context"
 	"time"
 
+	"github.com/bxcodec/go-clean-arch/domain"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/sync/errgroup"
-
-	"github.com/bxcodec/go-clean-arch/domain"
 )
-
-// ArticleRepository represent the article's repository contract
-//
-//go:generate mockery --name ArticleRepository
-type ArticleRepository interface {
-	Fetch(ctx context.Context, cursor string, num int64) (res []domain.Article, nextCursor string, err error)
-	GetByID(ctx context.Context, id int64) (domain.Article, error)
-	GetByTitle(ctx context.Context, title string) (domain.Article, error)
-	Update(ctx context.Context, ar *domain.Article) error
-	Store(ctx context.Context, a *domain.Article) error
-	Delete(ctx context.Context, id int64) error
-}
-
-// AuthorRepository represent the author's repository contract
-//
-//go:generate mockery --name AuthorRepository
-type AuthorRepository interface {
-	GetByID(ctx context.Context, id int64) (domain.Author, error)
-}
-
-type Service struct {
-	articleRepo ArticleRepository
-	authorRepo  AuthorRepository
-}
-
-// NewService will create a new article service object
-func NewService(a ArticleRepository, ar AuthorRepository) *Service {
-	return &Service{
-		articleRepo: a,
-		authorRepo:  ar,
-	}
-}
 
 /*
 * In this function below, I'm using errgroup with the pipeline pattern
