+++ service/service.go
@@ -1,0 +1,38 @@
+package service
+
+import (
+	"context"
+	"github.com/bxcodec/go-clean-arch/domain"
+)
+
+// ArticleRepository represent the article's repository contract
+//
+//go:generate mockery --name ArticleRepository
+type ArticleRepository interface {
+	Fetch(ctx context.Context, cursor string, num int64) (res []domain.Article, nextCursor string, err error)
+	GetByID(ctx context.Context, id int64) (domain.Article, error)
+	GetByTitle(ctx context.Context, title string) (domain.Article, error)
+	Update(ctx context.Context, ar *domain.Article) error
+	Store(ctx context.Context, a *domain.Article) error
+	Delete(ctx context.Context, id int64) error
+}
+
+// AuthorRepository represent the author's repository contract
+//
+//go:generate mockery --name AuthorRepository
+type AuthorRepository interface {
+	GetByID(ctx context.Context, id int64) (domain.Author, error)
+}
+
+type Service struct {
+	articleRepo ArticleRepository
+	authorRepo  AuthorRepository
+}
+
+// NewService will create a new article service object
+func NewService(a ArticleRepository, ar AuthorRepository) *Service {
+	return &Service{
+		articleRepo: a,
+		authorRepo:  ar,
+	}
+}
