+++ service/service_test.go
@@ -1,4 +1,4 @@
-package article_test
+package service_test
 
 import (
 	"context"
@@ -8,9 +8,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
-	"github.com/bxcodec/go-clean-arch/article"
-	"github.com/bxcodec/go-clean-arch/article/mocks"
 	"github.com/bxcodec/go-clean-arch/domain"
+	"github.com/bxcodec/go-clean-arch/service"
+	"github.com/bxcodec/go-clean-arch/service/mocks"
 )
 
 func TestFetchArticle(t *testing.T) {
@@ -32,7 +32,7 @@ func TestFetchArticle(t *testing.T) {
 		}
 		mockAuthorrepo := new(mocks.AuthorRepository)
 		mockAuthorrepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(mockAuthor, nil)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 		num := int64(1)
 		cursor := "12"
 		list, nextCursor, err := u.Fetch(context.TODO(), cursor, num)
@@ -51,7 +51,7 @@ func TestFetchArticle(t *testing.T) {
 			mock.AnythingOfType("int64")).Return(nil, "", errors.New("Unexpexted Error")).Once()
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 		num := int64(1)
 		cursor := "12"
 		list, nextCursor, err := u.Fetch(context.TODO(), cursor, num)
@@ -79,7 +79,7 @@ func TestGetByID(t *testing.T) {
 		mockArticleRepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(mockArticle, nil).Once()
 		mockAuthorrepo := new(mocks.AuthorRepository)
 		mockAuthorrepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(mockAuthor, nil)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		a, err := u.GetByID(context.TODO(), mockArticle.ID)
 
@@ -93,7 +93,7 @@ func TestGetByID(t *testing.T) {
 		mockArticleRepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(domain.Article{}, errors.New("Unexpected")).Once()
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		a, err := u.GetByID(context.TODO(), mockArticle.ID)
 
@@ -119,7 +119,7 @@ func TestStore(t *testing.T) {
 		mockArticleRepo.On("Store", mock.Anything, mock.AnythingOfType("*domain.Article")).Return(nil).Once()
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		err := u.Store(context.TODO(), &tempMockArticle)
 
@@ -137,7 +137,7 @@ func TestStore(t *testing.T) {
 		mockAuthorrepo := new(mocks.AuthorRepository)
 		mockAuthorrepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(mockAuthor, nil)
 
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		err := u.Store(context.TODO(), &mockArticle)
 
@@ -160,7 +160,7 @@ func TestDelete(t *testing.T) {
 		mockArticleRepo.On("Delete", mock.Anything, mock.AnythingOfType("int64")).Return(nil).Once()
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		err := u.Delete(context.TODO(), mockArticle.ID)
 
@@ -168,11 +168,11 @@ func TestDelete(t *testing.T) {
 		mockArticleRepo.AssertExpectations(t)
 		mockAuthorrepo.AssertExpectations(t)
 	})
-	t.Run("article-is-not-exist", func(t *testing.T) {
+	t.Run("service-is-not-exist", func(t *testing.T) {
 		mockArticleRepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(domain.Article{}, nil).Once()
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		err := u.Delete(context.TODO(), mockArticle.ID)
 
@@ -184,7 +184,7 @@ func TestDelete(t *testing.T) {
 		mockArticleRepo.On("GetByID", mock.Anything, mock.AnythingOfType("int64")).Return(domain.Article{}, errors.New("Unexpected Error")).Once()
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		err := u.Delete(context.TODO(), mockArticle.ID)
 
@@ -206,7 +206,7 @@ func TestUpdate(t *testing.T) {
 		mockArticleRepo.On("Update", mock.Anything, &mockArticle).Once().Return(nil)
 
 		mockAuthorrepo := new(mocks.AuthorRepository)
-		u := article.NewService(mockArticleRepo, mockAuthorrepo)
+		u := service.NewService(mockArticleRepo, mockAuthorrepo)
 
 		err := u.Update(context.TODO(), &mockArticle)
 		assert.NoError(t, err)
